Fix bug that caused files transmitted by a NetBSD box to have file names of (null).   CVS: ---------------------------------------------------------------------- automatically CVS: CVS: Committing in .  CVS: CVS: Modified Files: ----------------------------------------------------------------------
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@8922 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/bsd/krcp.c
@@ -748,7 +748,7 @@ void source(argc, argv)
 	    }
 	}
 	(void) sprintf(buf, "C%04o %ld %s\n",
-		       stb.st_mode&07777, stb.st_size, last);
+		       (int) stb.st_mode&07777, (long ) stb.st_size, last);
 	(void) des_write(rem, buf, strlen(buf));
 	if (response() < 0) {
 	    (void) close(f);
